# src/Blog.UI/api/swagger.json
{
"x-generator": "NSwag v12.0.15.0 (NJsonSchema v9.13.22.0 (Newtonsoft.Json v11.0.0.0))",
"swagger": "2.0",
"info": {
"title": "My Blog",
"description": "My personal blog",
"license": {
"name": "MIT",
"url": "https://github.com/PDmatrix/Blog/blob/master/LICENSE"
},
"version": "1.0"
},
"host": "localhost:5000",
"schemes": [
"http"
],
"paths": {
"/api/Posts": {
"get": {
"tags": [
"Posts"
],
"operationId": "Posts_GetAll",
"produces": [
"application/json"
],
"parameters": [
{
"type": "integer",
"name": "page",
"in": "query",
"format": "int32",
"default": 1,
"x-nullable": false
}
],
"responses": {
"200": {
"x-nullable": true,
"description": "",
"schema": {
"type": "array",
"items": {
"$ref": "#/definitions/PostPreviewDto"
}
}
},
"default": {
"x-nullable": true,
"description": "",
"schema": {
"$ref": "#/definitions/ProblemDetails"
}
}
},
"security": [
{
"JWT": []
}
]
},
"post": {
"tags": [
"Posts"
],
"operationId": "Posts_Create",
"produces": [
"application/json"
],
"parameters": [
{
"name": "postRequest",
"in": "body",
"required": true,
"schema": {
"$ref": "#/definitions/PostRequest"
},
"x-nullable": false
}
],
"responses": {
"201": {
"description": ""
},
"400": {
"x-nullable": true,
"description": "",
"schema": {
"$ref": "#/definitions/ProblemDetails"
}
},
"default": {
"x-nullable": true,
"description": "",
"schema": {
"$ref": "#/definitions/ProblemDetails"
}
}
},
"security": [
{
"JWT": []
}
]
}
},
"/api/Posts/{id}": {
"get": {
"tags": [
"Posts"
],
"operationId": "Posts_GetById",
"produces": [
"application/json"
],
"parameters": [
{
"type": "integer",
"name": "id",
"in": "path",
"required": true,
"format": "int32",
"x-nullable": false
}
],
"responses": {
"200": {
"x-nullable": true,
"description": "",
"schema": {
"$ref": "#/definitions/PostDto"
}
},
"404": {
"x-nullable": true,
"description": "",
"schema": {
"$ref": "#/definitions/ProblemDetails"
}
},
"default": {
"x-nullable": true,
"description": "",
"schema": {
"$ref": "#/definitions/ProblemDetails"
}
}
},
"security": [
{
"JWT": []
}
]
},
"delete": {
"tags": [
"Posts"
],
"operationId": "Posts_Delete",
"parameters": [
{
"type": "integer",
"name": "id",
"in": "path",
"required": true,
"format": "int32",
"x-nullable": false
}
],
"responses": {
"204": {
"description": ""
}
},
"security": [
{
"JWT": []
}
]
},
"put": {
"tags": [
"Posts"
],
"operationId": "Posts_Update",
"produces": [
"application/json"
],
"parameters": [
{
"type": "integer",
"name": "id",
"in": "path",
"required": true,
"format": "int32",
"x-nullable": false
},
{
"name": "postRequest",
"in": "body",
"required": true,
"schema": {
"$ref": "#/definitions/PostRequest"
},
"x-nullable": false
}
],
"responses": {
"204": {
"description": ""
},
"400": {
"x-nullable": true,
"description": "",
"schema": {
"$ref": "#/definitions/ProblemDetails"
}
},
"404": {
"x-nullable": true,
"description": "",
"schema": {
"$ref": "#/definitions/ProblemDetails"
}
},
"default": {
"x-nullable": true,
"description": "",
"schema": {
"$ref": "#/definitions/ProblemDetails"
}
}
},
"security": [
{
"JWT": []
}
]
}
},
"/api/Posts/preview": {
"post": {
"tags": [
"Posts"
],
"operationId": "Posts_Preview",
"produces": [
"application/json"
],
"parameters": [
{
"name": "previewRequest",
"in": "body",
"required": true,
"schema": {
"$ref": "#/definitions/PreviewRequest"
},
"x-nullable": false
}
],
"responses": {
"200": {
"x-nullable": true,
"description": "",
"schema": {
"$ref": "#/definitions/PreviewDto"
}
},
"default": {
"x-nullable": true,
"description": "",
"schema": {
"$ref": "#/definitions/ProblemDetails"
}
}
},
"security": [
{
"JWT": []
}
]
}
},
"/api/Admin/login": {
"post": {
"tags": [
"Admin"
],
"operationId": "Admin_Login",
"produces": [
"application/json"
],
"parameters": [
{
"name": "adminRequest",
"in": "body",
"required": true,
"schema": {
"$ref": "#/definitions/AdminRequest"
},
"x-nullable": false
}
],
"responses": {
"200": {
"x-nullable": true,
"description": "",
"schema": {
"type": "string"
}
},
"default": {
"x-nullable": true,
"description": "",
"schema": {
"$ref": "#/definitions/ProblemDetails"
}
}
},
"security": [
{
"JWT": []
}
]
}
}
},
"definitions": {
"ProblemDetails": {
"type": "object",
"description": "A machine-readable format for specifying errors in HTTP API responses based on https://tools.ietf.org/html/rfc7807.",
"additionalProperties": {},
"properties": {
"type": {
"type": "string",
"description": "A URI reference [RFC3986] that identifies the problem type. This specification encourages that, when\ndereferenced, it provide human-readable documentation for the problem type\n(e.g., using HTML [W3C.REC-html5-20141028]).  When this member is not present, its value is assumed to be\n\"about:blank\"."
},
"title": {
"type": "string",
"description": "A short, human-readable summary of the problem type.It SHOULD NOT change from occurrence to occurrence\nof the problem, except for purposes of localization(e.g., using proactive content negotiation;\nsee[RFC7231], Section 3.4)."
},
"status": {
"type": "integer",
"description": "The HTTP status code([RFC7231], Section 6) generated by the origin server for this occurrence of the problem.",
"format": "int32"
},
"detail": {
"type": "string",
"description": "A human-readable explanation specific to this occurrence of the problem."
},
"instance": {
"type": "string",
"description": "A URI reference that identifies the specific occurrence of the problem.It may or may not yield further information if dereferenced."
}
}
},
"PostPreviewDto": {
"type": "object",
"additionalProperties": false,
"required": [
"id"
],
"properties": {
"id": {
"type": "integer",
"format": "int32"
},
"title": {
"type": "string"
},
"excerpt": {
"type": "string"
}
}
},
"PostDto": {
"type": "object",
"additionalProperties": false,
"required": [
"id"
],
"properties": {
"id": {
"type": "integer",
"format": "int32"
},
"content": {
"type": "string"
},
"title": {
"type": "string"
}
}
},
"PostRequest": {
"type": "object",
"additionalProperties": false,
"properties": {
"content": {
"type": "string"
},
"title": {
"type": "string"
},
"excerpt": {
"type": "string"
}
}
},
"PreviewDto": {
"type": "object",
"additionalProperties": false,
"properties": {
"content": {
"type": "string"
}
}
},
"PreviewRequest": {
"type": "object",
"additionalProperties": false,
"properties": {
"content": {
"type": "string"
}
}
},
"AdminRequest": {
"type": "object",
"additionalProperties": false,
"properties": {
"user_name": {
"type": "string"
},
"password": {
"type": "string"
}
}
}
},
"securityDefinitions": {
"JWT": {
"type": "apiKey",
"description": "Type into the textbox: Bearer {your JWT token}.",
"name": "Authorization",
"in": "header"
}
}
}